+++ traces_api/database/model/unit.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime
+from sqlalchemy import Column, BigInteger, String, DateTime
 from traces_api.database import Base
 
 
@@ -6,7 +6,7 @@ class ModelUnit(Base):
 
     __tablename__ = 'unit'
 
-    id_unit = Column(Integer(), primary_key=True, autoincrement=True)
+    id_unit = Column(BigInteger(), primary_key=True, autoincrement=True)
     creation_time = Column(DateTime, nullable=False)
     last_update_time = Column(DateTime, nullable=False)
     annotation = Column(String(255))
